+++ wgsl_to_wgpu/src/lib.rs
@@ -448,7 +448,7 @@ fn vertex_input_structs(module: &naga::Module) -> Vec<TokenStream> {
                 quote! {
                     wgpu::VertexAttribute {
                         format: wgpu::VertexFormat::#format,
-                        offset: memoffset::offset_of!(#name, #field_name) as u64,
+                        offset: std::mem::offset_of!(#name, #field_name) as u64,
                         shader_location: #location,
                     }
                 }
@@ -660,22 +660,22 @@ mod test {
                     pub const VERTEX_ATTRIBUTES: [wgpu::VertexAttribute; 4] = [
                         wgpu::VertexAttribute {
                             format: wgpu::VertexFormat::Float32,
-                            offset: memoffset::offset_of!(VertexInput0, a) as u64,
+                            offset: std::mem::offset_of!(VertexInput0, a) as u64,
                             shader_location: 0,
                         },
                         wgpu::VertexAttribute {
                             format: wgpu::VertexFormat::Float32x2,
-                            offset: memoffset::offset_of!(VertexInput0, b) as u64,
+                            offset: std::mem::offset_of!(VertexInput0, b) as u64,
                             shader_location: 1,
                         },
                         wgpu::VertexAttribute {
                             format: wgpu::VertexFormat::Float32x3,
-                            offset: memoffset::offset_of!(VertexInput0, c) as u64,
+                            offset: std::mem::offset_of!(VertexInput0, c) as u64,
                             shader_location: 2,
                         },
                         wgpu::VertexAttribute {
                             format: wgpu::VertexFormat::Float32x4,
-                            offset: memoffset::offset_of!(VertexInput0, d) as u64,
+                            offset: std::mem::offset_of!(VertexInput0, d) as u64,
                             shader_location: 3,
                         },
                     ];
@@ -717,22 +717,22 @@ mod test {
                     pub const VERTEX_ATTRIBUTES: [wgpu::VertexAttribute; 4] = [
                         wgpu::VertexAttribute {
                             format: wgpu::VertexFormat::Float64,
-                            offset: memoffset::offset_of!(VertexInput0, a) as u64,
+                            offset: std::mem::offset_of!(VertexInput0, a) as u64,
                             shader_location: 0,
                         },
                         wgpu::VertexAttribute {
                             format: wgpu::VertexFormat::Float64x2,
-                            offset: memoffset::offset_of!(VertexInput0, b) as u64,
+                            offset: std::mem::offset_of!(VertexInput0, b) as u64,
                             shader_location: 1,
                         },
                         wgpu::VertexAttribute {
                             format: wgpu::VertexFormat::Float64x3,
-                            offset: memoffset::offset_of!(VertexInput0, c) as u64,
+                            offset: std::mem::offset_of!(VertexInput0, c) as u64,
                             shader_location: 2,
                         },
                         wgpu::VertexAttribute {
                             format: wgpu::VertexFormat::Float64x4,
-                            offset: memoffset::offset_of!(VertexInput0, d) as u64,
+                            offset: std::mem::offset_of!(VertexInput0, d) as u64,
                             shader_location: 3,
                         },
                     ];
@@ -775,22 +775,22 @@ mod test {
                     pub const VERTEX_ATTRIBUTES: [wgpu::VertexAttribute; 4] = [
                         wgpu::VertexAttribute {
                             format: wgpu::VertexFormat::Sint32,
-                            offset: memoffset::offset_of!(VertexInput0, a) as u64,
+                            offset: std::mem::offset_of!(VertexInput0, a) as u64,
                             shader_location: 0,
                         },
                         wgpu::VertexAttribute {
                             format: wgpu::VertexFormat::Sint32x2,
-                            offset: memoffset::offset_of!(VertexInput0, a) as u64,
+                            offset: std::mem::offset_of!(VertexInput0, a) as u64,
                             shader_location: 1,
                         },
                         wgpu::VertexAttribute {
                             format: wgpu::VertexFormat::Sint32x3,
-                            offset: memoffset::offset_of!(VertexInput0, a) as u64,
+                            offset: std::mem::offset_of!(VertexInput0, a) as u64,
                             shader_location: 2,
                         },
                         wgpu::VertexAttribute {
                             format: wgpu::VertexFormat::Sint32x4,
-                            offset: memoffset::offset_of!(VertexInput0, a) as u64,
+                            offset: std::mem::offset_of!(VertexInput0, a) as u64,
                             shader_location: 3,
                         },
                     ];
@@ -832,22 +832,22 @@ mod test {
                     pub const VERTEX_ATTRIBUTES: [wgpu::VertexAttribute; 4] = [
                         wgpu::VertexAttribute {
                             format: wgpu::VertexFormat::Uint32,
-                            offset: memoffset::offset_of!(VertexInput0, a) as u64,
+                            offset: std::mem::offset_of!(VertexInput0, a) as u64,
                             shader_location: 0,
                         },
                         wgpu::VertexAttribute {
                             format: wgpu::VertexFormat::Uint32x2,
-                            offset: memoffset::offset_of!(VertexInput0, b) as u64,
+                            offset: std::mem::offset_of!(VertexInput0, b) as u64,
                             shader_location: 1,
                         },
                         wgpu::VertexAttribute {
                             format: wgpu::VertexFormat::Uint32x3,
-                            offset: memoffset::offset_of!(VertexInput0, c) as u64,
+                            offset: std::mem::offset_of!(VertexInput0, c) as u64,
                             shader_location: 2,
                         },
                         wgpu::VertexAttribute {
                             format: wgpu::VertexFormat::Uint32x4,
-                            offset: memoffset::offset_of!(VertexInput0, d) as u64,
+                            offset: std::mem::offset_of!(VertexInput0, d) as u64,
                             shader_location: 3,
                         },
                     ];